Rename folder to interfaces.
