+++ SU2_CFD/src/interfaces/cfd/CConservativeVarsInterface.cpp
@@ -36,7 +36,7 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../../../include/transfer/cfd/CConservativeVarsInterface.hpp"
+#include "../../../include/interfaces/cfd/CConservativeVarsInterface.hpp"
 
 CConservativeVarsInterface::CConservativeVarsInterface(void) : CInterface() {
