+++ SU2_CFD/src/interfaces/cht/CConjugateHeatInterface.cpp
@@ -36,7 +36,7 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../../../include/transfer/cht/CConjugateHeatInterface.hpp"
+#include "../../../include/interfaces/cht/CConjugateHeatInterface.hpp"
 
 CConjugateHeatInterface::CConjugateHeatInterface(void) : CInterface() {
